[cli] split readline and stdio handler
This commit splits readline and stdio CLI handlers in POSIX app to
better demonstrate how to work with the CLI module.
